Refactored logic around skip invocation
